# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SiAP-PUSBANGKOMSDACKPS API
  version: 1.0.0
  description: API for SiAP-PUSBANGKOMSDACKPS (Sistem Aplikasi Pencairan - PUSBANGKOMSDACKPS)
servers:
  - url: /api/v1
    description: Initial API v1
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    DipaDokumenItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tahun_anggaran:
          type: integer
        bulan:
          type: integer
        revisi:
          type: integer
        file_hash_sha256:
          type: string
        original_name:
          type: string
        mime_type:
          type: string
        file_size_bytes:
          type: integer
        created_at:
          type: string
          format: date-time
        uploaded_by:
          type: string
          format: uuid
        uploaded_by_name:
          type: string
...
security:
  - cookieAuth: []
paths:
...
  /anggaran/upload-bukti:
    post:
      operationId: uploadBuktiAnggaran
      summary: Upload a document as proof for an anggaran node
      tags: [Anggaran]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                node_id:
                  type: string
                  format: uuid
              required:
                - file
                - node_id
      responses:
        '200':
          description: Document uploaded successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  /documents/{id}:
    get:
      operationId: downloadDocument
      summary: Download a document by ID
      tags: [Documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File content
        '404':
          description: Not found
    put:
      operationId: updateDocumentName
      summary: Update the name of a document
      tags: [Documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                original_name:
                  type: string
              required: [original_name]
      responses:
        '200':
          description: Document updated successfully
        '400':
          description: Bad request
    delete:
      operationId: deleteDocument
      summary: Delete a document by ID
      tags: [Documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document deleted successfully
        '404':
          description: Not found
  /anggaran/dipa/upload:
    post:
      operationId: uploadDipaDokumen
      summary: Upload a DIPA/RKKS PDF document
      tags: [Anggaran]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                tahun_anggaran:
                  type: integer
                bulan:
                  type: integer
                revisi:
                  type: integer
              required:
                - file
                - tahun_anggaran
                - bulan
                - revisi
      responses:
        '200':
          description: Document uploaded successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /anggaran/dipa/documents:
    get:
      operationId: getDipaDocuments
      summary: List DIPA/RKKS documents with optional filters
      tags: [Anggaran]
      parameters:
        - name: tahun
          in: query
          required: true
          schema:
            type: integer
        - name: bulan
          in: query
          required: false
          schema:
            type: integer
        - name: revisi
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of DIPA documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DipaDokumenItem'
  /anggaran/dipa/documents/{id}:
    delete:
      operationId: deleteDipaDokumen
      summary: Delete a DIPA/RKKS document by ID
      tags: [Anggaran]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document deleted successfully
        '404':
          description: Not found
...